Cleanup ID Attribute and add XML Docs
