+++ src/HotChocolate/Core/src/Types/Types/Relay/Extensions/RelayIdFieldExtensions.cs
@@ -1,22 +1,72 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using HotChocolate.Configuration;
-using HotChocolate.Internal;
-using HotChocolate.Types.Descriptors;
-using HotChocolate.Types.Descriptors.Definitions;
+using HotChocolate.Language;
 using HotChocolate.Types.Relay;
-using Microsoft.Extensions.DependencyInjection;
 
 #nullable enable
 
 namespace HotChocolate.Types
 {
+    /// <summary>
+    /// <c>.ID()</c> marks a field or parameter as a Global Unique Id.
+    /// The type of the target is rewritten to <c>ID</c> and a middleware is registered that,
+    /// automatically combines the value of fields annotated as ID with another value to form a
+    /// global identifier.
+    /// </summary>
+    /// <remarks>
+    /// Per default, this additional value is the name of the type the Id belongs to.
+    /// Since type names are unique within a schema, this ensures that we are returning a unique
+    /// Id within the schema. If our GraphQL server serves multiple schemas, the schema name
+    /// is also included in this combined Id. The resulting Id is then Base64 encoded to make
+    /// it opaque.
+    /// </remarks>
+    /// <example>
+    /// <para>
+    /// A field can be rewritten to a id by adding <c>[ID]</c> to the resolver.
+    /// </para>
+    /// <code>
+    /// public class User
+    /// {
+    ///     public int Id {get; set;}
+    /// }
+    /// public class UserType : ObjectType&gt;User>
+    /// {
+    ///     protected override void Configure(IObjectTypeDescriptor&gt;User> descriptor)
+    ///     {
+    ///         descriptor.Field(x => x.User).ID();
+    ///     }
+    /// }
+    /// </code>
+    /// <para>
+    /// In the resulting schema, the field `<c>User.id</c>` will be rewritten from `<c>Int</c>`
+    /// to `<c>ID</c>`
+    /// </para>
+    /// <code>
+    /// type User
+    /// {
+    ///     id: ID!
+    /// }
+    /// </code>
+    /// <para>
+    /// If `<c>User.id</c>` is requested in a query, the value is transformed to a base64 string
+    /// combined with the typename
+    /// Assuming `<c>User.id</c>` has the value 1. The following string is base64 encoded
+    /// <code>
+    /// User
+    /// i1
+    /// </code>
+    /// results in
+    /// <code>
+    /// VXNlcgppMQ==
+    /// </code>
+    /// </para>
+    /// </example>
     public static class RelayIdFieldExtensions
     {
-        private static IdSerializer? _idSerializer;
-
+        /// <inheritdoc cref="RelayIdFieldExtensions"/>
+        /// <param name="descriptor">the the descriptor</param>
+        /// <param name="typeName">
+        /// Sets the <see cref="IDAttribute.TypeName">type name</see> of the relay id
+        /// </param>
         public static IInputFieldDescriptor ID(
             this IInputFieldDescriptor descriptor,
             NameString typeName = default)
@@ -26,13 +76,16 @@ namespace HotChocolate.Types
                 throw new ArgumentNullException(nameof(descriptor));
             }
 
-            descriptor.Extend().OnBeforeCreate(RewriteInputFieldType);
-            descriptor.Extend().OnBeforeCompletion(
-                (c, d) => AddSerializerToInputField(c, d, typeName));
+            RelayIdFieldHelpers.ApplyIdToField(descriptor, typeName);
 
             return descriptor;
         }
 
+        /// <inheritdoc cref="RelayIdFieldExtensions"/>
+        /// <param name="descriptor">the the descriptor</param>
+        /// <param name="typeName">
+        /// Sets the <see cref="IDAttribute.TypeName">type name</see> of the relay id
+        /// </param>
         public static IArgumentDescriptor ID(
             this IArgumentDescriptor descriptor,
             NameString typeName = default)
@@ -42,13 +95,16 @@ namespace HotChocolate.Types
                 throw new ArgumentNullException(nameof(descriptor));
             }
 
-            descriptor.Extend().OnBeforeCreate(RewriteInputFieldType);
-            descriptor.Extend().OnBeforeCompletion(
-                (c, d) => AddSerializerToInputField(c, d, typeName));
+            RelayIdFieldHelpers.ApplyIdToField(descriptor, typeName);
 
             return descriptor;
         }
 
+        /// <inheritdoc cref="RelayIdFieldExtensions"/>
+        /// <param name="descriptor">the the descriptor</param>
+        /// <param name="typeName">
+        /// Sets the <see cref="IDAttribute.TypeName">type name</see> of the relay id
+        /// </param>
         public static IObjectFieldDescriptor ID(
             this IObjectFieldDescriptor descriptor,
             NameString typeName = default)
@@ -58,215 +114,23 @@ namespace HotChocolate.Types
                 throw new ArgumentNullException(nameof(descriptor));
             }
 
-            ResultConverterDefinition placeholder =
-                new((_, r) => r, key: WellKnownMiddleware.GlobalId, isRepeatable: false);
-
-            descriptor.Extend().Definition.ResultConverters.Add(placeholder);
-            descriptor.Extend().OnBeforeCreate(RewriteObjectFieldType);
-            descriptor.Extend().OnBeforeCompletion(
-                (c, d) => AddSerializerToObjectField(c, d, placeholder, typeName));
+            RelayIdFieldHelpers.ApplyIdToField(descriptor, typeName);
 
             return descriptor;
         }
 
-        public static IInterfaceFieldDescriptor ID(
-            this IInterfaceFieldDescriptor descriptor)
+        /// <inheritdoc cref="RelayIdFieldExtensions"/>
+        /// <param name="descriptor">the the descriptor</param>
+        public static IInterfaceFieldDescriptor ID(this IInterfaceFieldDescriptor descriptor)
         {
             if (descriptor is null)
             {
                 throw new ArgumentNullException(nameof(descriptor));
             }
 
-            descriptor.Extend().OnBeforeCreate(RewriteInterfaceFieldType);
+            RelayIdFieldHelpers.ApplyIdToField(descriptor);
 
             return descriptor;
-        }
-
-        private static void RewriteInputFieldType(
-            IDescriptorContext context,
-            ArgumentDefinition definition)
-        {
-            if(definition.Type is ExtendedTypeReference typeReference)
-            {
-                ITypeInfo typeInfo = context.TypeInspector.CreateTypeInfo(typeReference.Type);
-                IExtendedType type = RewriteType(context.TypeInspector, typeInfo);
-                definition.Type = typeReference.WithType(type);
-            }
-            else
-            {
-                throw new SchemaException(SchemaErrorBuilder.New()
-                    .SetMessage("Unable to resolve type from field `{0}`.", definition.Name)
-                    .Build());
-            }
-        }
-
-        private static void RewriteObjectFieldType(
-            IDescriptorContext context,
-            ObjectFieldDefinition definition)
-        {
-            if(definition.Type is ExtendedTypeReference typeReference)
-            {
-                ITypeInfo typeInfo = context.TypeInspector.CreateTypeInfo(typeReference.Type);
-                IExtendedType type = RewriteType(context.TypeInspector, typeInfo);
-                definition.Type = typeReference.WithType(type);
-            }
-            else
-            {
-                throw new SchemaException(SchemaErrorBuilder.New()
-                    .SetMessage("Unable to resolve type from field `{0}`.", definition.Name)
-                    .Build());
-            }
-        }
-
-        private static void RewriteInterfaceFieldType(
-            IDescriptorContext context,
-            InterfaceFieldDefinition definition)
-        {
-            if(definition.Type is ExtendedTypeReference typeReference)
-            {
-                ITypeInfo typeInfo = context.TypeInspector.CreateTypeInfo(typeReference.Type);
-                IExtendedType type = RewriteType(context.TypeInspector, typeInfo);
-                definition.Type = typeReference.WithType(type);
-            }
-            else
-            {
-                throw new SchemaException(SchemaErrorBuilder.New()
-                    .SetMessage("Unable to resolve type from field `{0}`.", definition.Name)
-                    .Build());
-            }
-        }
-
-        private static IExtendedType RewriteType(ITypeInspector typeInspector, ITypeInfo typeInfo)
-        {
-            Type current = typeof(IdType);
-
-            if (typeInfo.Components.Count > 1)
-            {
-                foreach (TypeComponent component in typeInfo.Components.Reverse().Skip(1))
-                {
-                    if (component.Kind == TypeComponentKind.NonNull)
-                    {
-                        current = typeof(NonNullType<>).MakeGenericType(current);
-                    }
-                    else if (component.Kind == TypeComponentKind.List)
-                    {
-                        current = typeof(ListType<>).MakeGenericType(current);
-                    }
-                }
-            }
-
-            return typeInspector.GetType(current);
-        }
-
-        private static void AddSerializerToInputField(
-            ITypeCompletionContext completionContext,
-            ArgumentDefinition definition,
-            NameString typeName)
-        {
-            ITypeInspector typeInspector = completionContext.TypeInspector;
-            IExtendedType? resultType;
-
-            if (definition is InputFieldDefinition inputField)
-            {
-                resultType = typeInspector.GetReturnType(inputField.Property!, true);
-            }
-            else if (definition.Parameter is not null)
-            {
-                resultType = typeInspector.GetArgumentType(definition.Parameter, true);
-            }
-            else if (definition.Type is ExtendedTypeReference typeReference)
-            {
-                resultType = typeReference.Type;
-            }
-            else
-            {
-                throw new SchemaException(SchemaErrorBuilder.New()
-                    .SetMessage("Unable to resolve type from field `{0}`.", definition.Name)
-                    .SetTypeSystemObject(completionContext.Type)
-                    .Build());
-            }
-
-            definition.Formatters.Add(CreateSerializer(completionContext, resultType, typeName));
-        }
-
-        private static void AddSerializerToObjectField(
-            ITypeCompletionContext completionContext,
-            ObjectFieldDefinition definition,
-            ResultConverterDefinition placeholder,
-            NameString typeName)
-        {
-            ITypeInspector typeInspector = completionContext.TypeInspector;
-            IExtendedType? resultType;
-
-            if (definition.ResultType is not null)
-            {
-                resultType = typeInspector.GetType(definition.ResultType);
-            }
-            else if (definition.Type is ExtendedTypeReference typeReference)
-            {
-                resultType = typeReference.Type;
-            }
-            else
-            {
-                throw new SchemaException(SchemaErrorBuilder.New()
-                    .SetMessage("Unable to resolve type from field `{0}`.", definition.Name)
-                    .SetTypeSystemObject(completionContext.Type)
-                    .Build());
-            }
-
-            NameString schemaName = default;
-            completionContext.DescriptorContext.SchemaCompleted += (_, args) =>
-                schemaName = args.Schema.Name;
-
-            IIdSerializer serializer =
-                completionContext.Services.GetService<IIdSerializer>() ??
-                new IdSerializer();
-            var index = definition.ResultConverters.IndexOf(placeholder);
-
-            if (typeName.IsEmpty)
-            {
-                typeName = completionContext.Type.Name;
-            }
-
-            definition.ResultConverters[index] = new((_, result) =>
-            {
-                if (result is not null)
-                {
-                    if (resultType.IsArrayOrList)
-                    {
-                        var list = new List<object?>();
-
-                        foreach (var element in (IEnumerable)result)
-                        {
-                            list.Add(element is null
-                                ? element
-                                : serializer.Serialize(schemaName, typeName, element));
-                        }
-
-                        return list;
-                    }
-
-                    return serializer.Serialize(schemaName, typeName, result);
-                }
-
-                return result;
-            }, key: WellKnownMiddleware.GlobalId, isRepeatable: false);
-        }
-
-        private static IInputValueFormatter CreateSerializer(
-            ITypeCompletionContext completionContext,
-            IExtendedType resultType,
-            NameString typeName)
-        {
-            IIdSerializer serializer =
-                completionContext.Services.GetService<IIdSerializer>() ??
-                (_idSerializer ??= new IdSerializer());
-
-            return new GlobalIdInputValueFormatter(
-                typeName.HasValue ? typeName : completionContext.Type.Name,
-                serializer,
-                resultType,
-                typeName.HasValue);
         }
     }
 }
